# swagger.yaml
openapi: 3.0.0
info:
  title: twitterGram
  description: A social media API influenced greatly by twitter, built with node, typescript, express, sequelize and a lot more.
  contact: {}
  version: "1.0"
servers:
  - url: https://twitter-gram-api.herokuapp.com/api/v1
    variables: {}
paths:
...
  /posts:
    post:
      tags:
        - Post
      summary: create post
      description: You can create post with this route
      operationId: createpost
      parameters:
        - name: x-refresh
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example: "{{refreshToken}}"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                caption:
                  type: string
                file:
                  type: string
                  format: binary
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    get:
      tags:
        - Post
      summary: get feeds
      description: This gets the post of those the user follows. user must follow other's to see their posts
      operationId: getfeeds
      parameters:
        - name: x-refresh
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example: "{{refreshToken}}"
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
  /posts/{postId}:
    parameters:
      - in: path
        name: postId
        schema:
          type: integer
        required: true
        description: The post ID
    get:
      tags:
        - Post
      summary: get a post
      description: Returns a particular post given the post id as params
      operationId: getapost
      parameters:
        - name: x-refresh
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example: "{{refreshToken}}"
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    delete:
      tags:
        - Post
      summary: delete post
      description: >-
        Use this route to delete posts given the post id as params


        NB: Only the owner of the post can delete the post
      operationId: deletepost
      parameters:
        - name: x-refresh
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example: "{{refreshToken}}"
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    patch:
      tags:
        - Post
      summary: update post
      description: >-
        This route is used to update user's post, it requires the post id and the new post content.

        NB: Only the owner of the post can update the post
      operationId: updatepost
      parameters:
        - name: x-refresh
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example: "{{refreshToken}}"
        - name: postId
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
  /comments:
    post:
      tags:
        - Comment
      summary: create comment
      description: This route is used to comment on a particular post
      operationId: createcomment
      parameters:
        - name: x-refresh
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example: "{{refreshToken}}"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                text:
                  type: string
                file:
                  type: string
                  format: binary
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /comments/{commentId}:
    parameters:
      - in: path
        name: commentId
        schema:
          type: integer
        required: true
        description: The post ID
    patch:
      tags:
        - Comment
      summary: edit comment
      description: >-
        User's comments are updated with this route.

        Again, only the user that created the comment can update it.
      operationId: editcomment
      parameters:
        - name: x-refresh
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example: "{{refreshToken}}"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false

    delete:
      tags:
        - Comment
      summary: delete comment
      description: This route is used to delete a comment.
      operationId: deletecomment
      parameters:
        - name: x-refresh
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example: "{{refreshToken}}"
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
  /posts/image/post_ouxrycvpzy1638738116213.jpg:
    get:
      tags:
        - Image
      summary: get image
      description: Get an image
      operationId: getimage
      parameters:
        - name: x-refresh
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example: "{{refreshToken}}"
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
components:
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...